rework import semantics, test structure
just realized you didn't need to preface imported modules with `super`

also tests should be organized by their module. They probably shouldn't
all be in the same file
