+++ src/actor/test.rs
@@ -1,0 +1,9 @@
+#[cfg(test)]
+mod test {
+
+    #[test]
+    fn actor_test() {
+        panic!("fdslkjsdlkjf");
+    }
+
+}
